git-svn-id: http://kos.informatik.uni-osnabrueck.de/software/meshing@2391 bea721bf-f9f1-0310-a2d5-e0bfdc8dd81e

+++ src/lib3d/PlyStructs.h
@@ -40,18 +40,18 @@ struct PlyFaceDescription{
 };
 
 struct PlyVertex{
-  double x;
-  double y;
-  double z;
-  double nx;
-  double ny;
-  double nz;
+  float x;
+  float y;
+  float z;
+  float nx;
+  float ny;
+  float nz;
   float r;
   float g;
   float b;
 
-  double u;
-  double v;
+  float u;
+  float v;
 
   int texture;
 };
